// frontend/src/backend.d.ts
export type Time = bigint;
export interface QuestProgress {
    user: Principal;
    completed: boolean;
    questId: bigint;
    completedSteps: Array<bigint>;
}
export type WalletOperationResult = {
    __kind__: "error";
    error: WalletError;
} | {
    __kind__: "success";
    success: Array<EpiqWallet>;
};
export interface QuizQuestion {
    question: string;
    correctAnswer: bigint;
    options: Array<string>;
}
export interface CourseModuleContent {
    doStep?: Array<QuizQuestion>;
    show?: string;
    tell: string;
}
export interface CourseWithModules {
    id: bigint;
    reward: bigint;
    title: string;
    availableToAll: boolean;
    description: string;
    assignedMemberTypes: Array<MemberType>;
    category: string;
    modules: Array<CourseModuleWithQuestions>;
}
export interface Transaction {
    id: bigint;
    to: Principal;
    status: string;
    from: Principal;
    currency: string;
    timestamp: Time;
    amount: bigint;
}
export interface CourseModule {
    id: bigint;
    title: string;
    description: string;
    steps: Array<string>;
    category: string;
}
// ...
export interface EpiqWallet {
    publicAddress: string;
    metadata?: string;
    chain: Chain;
    createdAt: Time;
    walletLabel?: string;
    walletId: bigint;
}
export interface AuthenticationEvent {
    status: string;
    epiqId?: string;
    failureReason?: string;
    authMethod: string;
    ipContext?: string;
    timestamp: Time;
    principalId: Principal;
}
export interface EpqReward {
    id: bigint;
    verified: boolean;
    source: string;
    user: Principal;
    timestamp: Time;
    amount: bigint;
}
export interface QuestModule {
    id: bigint;
    reward: bigint;
    title: string;
    description: string;
    steps: Array<QuestStep>;
    category: string;
}
export type WalletOperationResult_1 = {
    __kind__: "error";
    error: WalletError;
} | {
    __kind__: "success";
    success: WalletCreationResult;
};
export interface QuestProgressRecord {
    moduleId: bigint;
    user: Principal;
    completed: boolean;
    completedSteps: Array<bigint>;
    timestamp: Time;
}
export interface http_header {
    value: string;
    name: string;
}
export interface http_request_result {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface CourseModuleWithQuestions {
    id: bigint;
    title: string;
    content: CourseModuleContent;
    description: string;
    questions: Array<QuizQuestion>;
}
export interface WalletCreationResult {
    chain: Chain;
    privateKey: string;
    address: string;
    recoveryPhrase: string;
}
// ...
export interface DisplayNameChangeHistory {
    timestamp: Time;
    principalId?: Principal;
    newName: string;
    oldName: string;
}
export interface UserProfile {
    icpWalletAddress: string;
    displayName: string;
    name: string;
    email: string;
    importedFromFirebase: boolean;
    memberType: string;
    btcWalletAddress: string;
    ethWalletAddress: string;
}
export enum Chain {
    btc = "btc",
    evm = "evm",
    icp = "icp",
    custom = "custom"
}
export enum MemberType {
    member = "member",
    trustee = "trustee",
    admin = "admin",
    citizen = "citizen",
    beneficiary = "beneficiary",
    superAdmin = "superAdmin",
    business = "business",
    ambassador = "ambassador",
    partner = "partner"
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}
export enum WalletError {
    invalidRequest = "invalidRequest",
    unauthorizedAccess = "unauthorizedAccess",
    internalError = "internalError",
    walletLimitReached = "walletLimitReached"
}